Climber Limits Idea
Just an extra if-else statement that should be all that we need to limit the vertical movement of the climber. The function to get the position of the motor is defiantly wrong, but it is just there as a place holder. I just made this in my spare time to get more practice with GitHub and Java.